moved function to same package
