+++ packages/workspace/src/workspace/index_utils.ts
@@ -1,0 +1,28 @@
+/*
+*                      Copyright 2022 Salto Labs Ltd.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with
+* the License.  You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+import { ReadOnlyElementsSource, Element, Change, toChange } from '@salto-io/adapter-api'
+import { collections } from '@salto-io/lowerdash'
+
+const { awu } = collections.asynciterable
+
+export const getAllElementsChanges = async (
+  currentChanges: Change<Element>[],
+  elementsSource: ReadOnlyElementsSource
+): Promise<Change<Element>[]> =>
+  awu(await elementsSource.getAll())
+    .map(element => toChange({ after: element }))
+    .concat(currentChanges)
+    .toArray()
